requirement.txt
streamlit
google-generativeai
pypdf
langchain
langchain-community
langchain-text-splitters
faiss-cpu